+++ NEWS.md
@@ -1,7 +1,8 @@
-# crew.aws.batch 0.0.6.9002 (development)
+# crew.aws.batch 0.0.6.9003 (development)
 
 * Send both cancellation and termination requests to end jobs.
 * Fix launcher bug/typo where parameters were supplied to container overrides.
+* Add a new `all` argument to `terminate()` in the AWS Batch monitor.
 
 # crew.aws.batch 0.0.6